This script package was developed by Ray Wang and David Kim.
The main script is pdb2vall.py and can be used to generate a VALL
fragment_picker template database from a specified PDB+chain.

There are a number of dependencies which are explained in the README.

Currently pdb2vall.py will not add actual values for bfactor, nali, 
CB, and centroid coordinates (placeholder values are used).




git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/rosetta_tools@48044 6189c7af-7406-0410-a2ce-f22f51069c23