Adding nice color output to configure_rosetta_repo.sh
The script has been tested and works on my machine.  One downside
has been observed in that the user's GitHub password must be pro-
vided for each file that is downloaded in addition to having SSH
keys set up for GitHub.  This seems like a lot and I am hoping to
find a solution soon.