Adding aliases and colors to the configuration script.
The structure of the script is no longer a while loop with a break
at the end, which is, duh, better.  I think this now does all of
the things we want it to do.

A modification I can forsee is that we may be moving the git_hooks
directory and that will need to be reflected here.