Changing the way aliases are set up.
In the configrue script, the alias set up worked fine in zsh, but
not in bash or sh.  Since the script is executed in sh, that's a
problem.  This commit addresses this issue.  The configuration has
been tested on OS X and Linux, but the aliases themselves were
only tested under OS X.  If someone wants to investigate further,
please do.