Append rather than override flags in coding_util/generate_warnings_list.py
This change allows generate_warnings_list.py to work on Macs running OS X
10.9 by not overriding the ftemplate-depth flag and UNUSUAL_ALLOCATOR
macro. This change also fixes a bug that made getting the warnings in
release mode impossible -- debug mode was built instead.
Finally, the warning level for clang was reduced to Wall+pedantic instead
of Weverything.  Hopefully this will result in a manageable number of warnings.