Reworked separation into first and second half
Light and heavy antibody sequences alike are split
into multiple conserved (or not so conserved) regions.
Today, the motifs for H1 (L1) and H3 (L3) are
determined on the sequence and the remnant locations,
e.g. H2 (L2), derived from those matches.

The motifs are not search in the whole sequence
but the sequence is first split into a first and second
part. This patch proposes to have that split dependent
on the position of the first match to thus accomodate
also for shifts in the location of the loops in the
user-provided sequence. We observed in reverse-translated
sequences from B cells.