Rewriter now handles cases where there are multiple "roots" to the element tree
e.g. if you have a whole bunch of movers and filters and score functions
you use in several scripts, then you might create an XML file that
does not have a ROSETTASCRIPTS element at the top, but still does have
a MOVERS element, a SCOREFXNS element, and a FILTERS element. These three
elements would be turned into roots for their own subtrees (they are closed,
after all) so there would not be a single root for the file, but instead 3.

The example_w_multiple_roots.xml file illustrates this case.