Fixing two issues with the beautifier
1. The beautifier now handles template specialization without the \<\>s such as now
found in core/id/NamedAtomID.hh

     template NamedAtomID::NamedAtomID(std::string const &, Size);

where no "<" is ever found. The beautifier now handles this appropriately. Previously,
it would keep reading until it found a "\<" and that might not happen, or it might
happen much lower in the file.

2. The beautifier never properly handled multi-line-template-parameter-list
indentation and so all the lines in the template parameters were flush to the
left, even if the original line was indented. Now all lines after the first
are indented one more level.