Made epitope_predictor_matrix.py look harder for the db
Added the capability to look directly for the Rosetta database, assuming
that this is being run from Rosetta/tools/mhc_energy_tools and there
also exists Rosetta/main/database.

Also changed os.environ to os.getenv, which returns None instead of
crashing if $ROSETTA hasn't been set, and tweaked the error handling.

The logic of how to find the file could maybe be cleaner, but this works
for now.