Fix beautifier handling of initializer-lists inside initializer lists
Previously, if a constructor's initializer list included the C++11
initializer list {} synatx, it would encounter the left curly brace
("{") and interpret it as the start of the constructor's body.

Now it is smarter.

Fixing a related problem where template-specialization functions
where the word "template" shows up but isn't immediately
followed by an opening left angle brace ("<").

e.g.

template void somefunc<int>(int a, int b);

Previously the beautifier would process everything
from "template" to the right angle brace as the template
part, and then hand the opening left paren to be interpretted
as a function. This doesn't work in the new scheme, as the opening
left paren needs to be paid attention to in
process_function_preamble_or_variable.

Now, if a non "<" is hit after the word template, then the process_template
function returns.