mhc_gen_db.py: Ensure that NATAA gets written to resfile for WT-only positions
When writing resfiles using mhc_gen_db.py, a PIKAA line is written for
any residue that allows more than one residue type (wild-type plus at
least one more).  It previously skipped WT-only residues (i.e. they were not
included in the resfile).  Now, if it is a WT-only residue, a NATAA line
will be written out.  (We could use PIKAA with only one residue, but
NATAA makes it clearer that we are dealing with the WT residue for the
user.)