mhc_gen_db.py: Added option to "offset" first residue in PSSM->resfile operations
If the user generates a PSSM from a PDB file that doesn't start with
residue 1, mhc_gen_db.py would assume that the first residue should be
1, and will fail due to a mismatch between the PSSM sequence and PDB
sequence.  The user can now indicate the residue number of the first
residue in the PSSM to apply an offset.  This also fixes the resfile
output.