mhc_energy_tools/epilib/expt_data.py: fix for bytearray issue
My previous fix didn't actually fix the issue where the mysql connector
was returning bytearrays instead of strings.  This seems to be platform
dependent, as Chris wasn't seeing it on his Mac.

I'm changing it so that the variables returned are decoded if they are
bytearrays, but not if they are not.  Done on a variable-by-variable
basis to be safe.  This should therefore not cause problems regardless
of what data type is returned.