mhc_energy_tools: Added demo-diff.sh to "smartly" check out directory output
After running all of the demos, the resulting files in the out directory
can be checked to see if they match the ref-out directory.  This is a
bit challenging, as some of the files are binary files.  This script
first diffs appropriate text files, and then checks the binary .db
files by dump contents to a text file and then diffing.

The idea here is that this can be treated a bit like an integration test
for the mhc_energy_tools suite.