epilib/expt_data.py: Added explicit error handling for mysql
For the os.system calls, if the shell returns a non-zero error code, it
prints an error message and python will exit, returning an error code back to the
shell.

For mysql.connector, the commands associated with accessing the database
are wrapped in a try/except block.  In the except case, the script will
terminate with an error code.