Add diff5.py, a tool for helping deduce merge conflicts.
For significant merge conflicts, I've found it helpful to see what has changed in each side of the merge.

Turning on merge.conflictstyle=diff3 in git helps, in that it gives you most recent common ancestor version in the diff,
but for large, complex diffs it's sometimes hard to parse (especially if there's a small change in a section of code which was
moved/deleted in the other leg of the merge.)

This diff5 tool takes a diff3 merge result and then uses some python diff tools to do add a line-by-line annotation of the
two legs of the diff, showing what has changed in each since their common merge base.