Remove includes from template files w surrogates
The surrogate file, which uses the functions defined in the
templated file, will be tested along with the tempalted file.

A header X.hh can be removed from template file A.hh that's
in turn included by surrogate file B.cc can be removed safely
if
1) A.hh compiles, and
2) B.hh compiles.

If removing X.hh disrupts the compilation of either file,
then it cannot be removed. It is insufficient, however, to remove
header X.hh from A.hh if the compilation of only one of the two
files is tested.

Also restructuring the code in remove_headers_from_all_files.py
and extending find_heavy_paths.py to consider the paths from
multiple starting nodes.