Refactor trim_unnecessary_headers_from_file to support sugraph include trimming
Add a new script for trimming a the includes from a subgraph of the inclusion
DAG. This will only attempt to remove includes from the files that
include/transclude a set of starting nodes and will only try to remove from
these files the includes that are included/transcluded by the starting
nodes. It should allow us to rapidly purge a subset of includes from the tree
without needing to wait 2 days for the whole_shebang script to run.