Fix transclude addition in subgraph cleanup script
Previously, transclusions of the starting nodes were added
to all of the files that transcluded the starting file EXCEPT
for the starting nodes themselves. This meant that when later
trying to remove files that were transcluded by the starting
node, the logic for figuring out whether the transclude could
be removed would say that it couldn't. The logic was:

consider removal of X_hh from C_cc if there is some starting_node
s.t.
1. starting_node transcludes X_hh, and
2. C_cc trancludes starting_node

Condition 1 was evaluated with the original transclusion graph,
but Condition 2 was evaluated with the current transclusion graph.
The second condition was disrupted with the transclusion-addition
error.

If C_cc contained the explicit inclusion of all the implicitly
transcluded files, then when the consideration of X_hh's removal
occurs, then the second condition would have been true. However,
since C_cc didn't get its transclusion of the starting nodes,
the second condition would evaluate to False and X_hh's removal
would be considered prohibited.